requirements.txt
pycryptodome
python-gnupg
PyYAML
